Updates Challenge Authorize to support 'processing' challange status
The RFC indicates that challenge statuses can be one of:  “pending”, “processing”, “valid”, and “invalid". Currently the code is only looking for "pending", "valid" or "invalid" statuses though. See spec reference:

- https://ietf-wg-acme.github.io/acme/draft-ietf-acme-acme.html#rfc.section.7.1.5
- https://ietf-wg-acme.github.io/acme/draft-ietf-acme-acme.html#identifier-validation-challenges

I know it's a minor thing but my internal custom CA is returning "processing" causing the certificate request to fail with: 

```
 [EROR] Authorization result: processing
 [EROR] Create certificate failed: Authorization failed
```

Maybe I'm missing something, but shouldn't "processing" be an available status for a challenge? The change seems pretty straightforward so I thought I'd just make a quick pull rather than file a bug but I'm happy to do so if you'd prefer!