Create ImportRDSFull.ps1
For a Remote Desktop server, the current powershell scripts do not take care of the distribution of the certificate to installed roles of  RD WebAccess, RD Redirector,  RD Gateway and RD Connection Broker.
The script is based on ImportRDS.ps1 and extends it.
In order for this script to work, the private key of the certificate needs to be marked as exportable. Set "PrivateKeyExportable" in settings.json to true.