Remove Duplicate Command From ImportRDSFull.ps1
An identical wmic command was included at the end of the try block for the RD Gateway cert configuration in addition to being the only active command in the subsequent RDP Listener try block.  Presumably the command should only be used in the latter try block so that its failure won't return the wrong error in the earlier try block.